implement post deletion

+++ resources/views/posts/edit.blade.php
@@ -15,9 +15,11 @@
 
         <div class="panel-footer text-right">
         	<!-- Delete -->
-            {!! Form::open(['route' => ['posts.destroy', $post], 'method' => 'delete']) !!}
-                <button type="submit" class="btn btn-link" title="Delete Post" data-toggle="tooltip" data-placement="bottom" style="padding: 0;">Delete</button>
-            {!! Form::close() !!}
+            @if (Auth::user() == $post->user || Auth::user()->admin)
+                {!! Form::open(['route' => ['posts.destroy', $post], 'method' => 'delete']) !!}
+                    <button type="submit" class="btn btn-link" title="Delete Post" data-toggle="tooltip" data-placement="bottom" style="padding: 0;">Delete</button>
+                {!! Form::close() !!}
+            @endif
         </div>
     </div>
 </div>
